test(compliance): pin v4 manifest additive-shape guarantee against frozen v3 (P3.2 AC-3)
The existing bump gate (test_the_shape_change_forces_the_bump_it_obliges) only
proves a bump of adequate SIZE happened -- a major-bumped BREAKING change would
satisfy it too. AC-3 (corrected 2026-09-07 after PR #686's Stage-1 REJECT) asks
for the narrower, actually-verifiable claim: the v4 manifest stays readable by
a v3-shaped consumer, additive nodes only, verified against a frozen v3
fixture. Pins that directly: diff.removed == [] and diff.retyped == [] between
the frozen v3.0 fixture and the live v4 contract.

Split test_traceability_contract.py's identity/integrity tests (duplicate-id
refusal, key<->id agreement, tombstone ordering) into a new sibling file to
keep both under the 300-LOC bloat-baseline threshold -- same subject family,
not a different one.
