+++ plugins/shipwright-compliance/tests/test_traceability_contract.py
@@ -1,4 +1,4 @@
-"""CONTRACT GATE for the test-traceability manifest -- schema v3 (campaign S3).
+"""CONTRACT GATE for the test-traceability manifest -- schema v3/v4 (campaign S3/P3.2).
 
 Pins the WIRE SHAPE and the key form across versions. The fixture repo, the git-baseline
 machinery and the constants live in ``traceability_contract_support``; the rationale for
@@ -23,13 +23,6 @@ _HERE = Path(__file__).resolve().parent
 if str(_HERE) not in sys.path:
     sys.path.insert(0, str(_HERE))
 
-from scripts.lib.collectors._test_links_requirements import (
-    DuplicateRequirementId,
-    KeyNotDerivedFromId,
-    ManifestIntegrityError,
-    assert_keys_derive_from_ids,
-    build_requirement_index,
-)
 from traceability_contract_support import (
     ARTIFACT,
     CB,
@@ -43,8 +36,6 @@ from traceability_contract_support import (
     _manifest_for,
     _materialize,
     _REPO_ROOT,
-    _SPEC,
-    _TEST,
     fixture_path,
     live_contract,
     published_anywhere,
@@ -135,6 +126,24 @@ class TestTheGate:
         )
         assert diff is not None
 
+    def test_v4_stays_additive_over_the_frozen_v3_shape(self, repo: Path):
+        """AC-3 (P3.2, corrected 2026-09-07): the v4 manifest must stay readable by a
+        v3-shaped consumer -- additive nodes only, nothing removed or retyped.
+
+        `test_the_shape_change_forces_the_bump_it_obliges` only proves a bump of
+        adequate SIZE happened; `required_bump` returns "major" for a BREAKING diff
+        too, so a major-bumped breaking change would satisfy that test as well. This
+        pins the narrower, stronger claim against the frozen v3.0 fixture directly --
+        not against whatever `origin/main` happens to currently publish -- because the
+        promise is about v3 shape survival specifically, not about the immediately
+        preceding version.
+        """
+        v3 = json.loads(fixture_path("3.0").read_text(encoding="utf-8"))
+        diff = CE.diff_skeletons(v3["contract"], live_contract(repo))
+        assert diff.removed == [], f"v3 fields dropped in v4: {diff.removed}"
+        assert diff.retyped == [], f"v3 fields retyped in v4: {diff.retyped}"
+        assert not diff.is_breaking
+
 
 class TestLoadBearingFields:
     def test_every_documented_field_is_on_the_wire(self, repo: Path):
@@ -182,108 +191,7 @@ class TestKeyFormIsIdDerived:
         assert node["spec_path"] == ".shipwright/planning/99-totally-renamed/spec.md"
 
 
-class TestDuplicateIdsFailClosed:
-    """A v3 key is a pure function of the id, so two specs CAN claim one key. Resolving
-    that by keeping either node would silently delete a requirement from the artifact
-    whose job is to reveal traceability gaps, so generation refuses instead."""
-
-    def test_two_specs_claiming_one_id_raise_and_name_both(self, tmp_path: Path):
-        _materialize(tmp_path, "01-a")
-        second = tmp_path / ".shipwright" / "planning" / "02-b" / "spec.md"
-        second.parent.mkdir(parents=True, exist_ok=True)
-        second.write_text(_SPEC, encoding="utf-8")
-        with pytest.raises(DuplicateRequirementId) as excinfo:
-            _manifest_for(tmp_path)
-        message = str(excinfo.value)
-        assert "FR-03.01" in message
-        # Actionable or it is not a usable error: it must name BOTH contributing specs.
-        assert "01-a" in message and "02-b" in message
-
-    def test_two_rows_in_ONE_spec_sharing_an_id_also_raise(self, tmp_path: Path):
-        """v2 collapsed this silently too (same namespace ⇒ same key), so v3 is not
-        inventing a failure here — it is making an already-silent loss visible."""
-        spec = tmp_path / ".shipwright" / "planning" / "01-a" / "spec.md"
-        spec.parent.mkdir(parents=True, exist_ok=True)
-        spec.write_text(_SPEC.replace("| FR-03.02 | Reporting rollup | Should | int, db |",
-                                      "| FR-03.01 | Duplicated row | Should | unit |"),
-                        encoding="utf-8")
-        (tmp_path / "tests").mkdir(parents=True, exist_ok=True)
-        (tmp_path / "tests" / "test_auth.py").write_text(_TEST, encoding="utf-8")
-        with pytest.raises(DuplicateRequirementId):
-            _manifest_for(tmp_path)
-
-
-class TestKeyAgreesWithItsNodeId:
-    """The schema pins the key SHAPE; only this pins that the two halves AGREE."""
-
-    def test_a_numerically_named_directory_cannot_smuggle_a_path_namespace(self):
-        """The regex alone would wave ``02::FR-03.01`` through — a repo whose split
-        directories are numbered (``02/``) is exactly where a reintroduced path-derived
-        namespace would look plausible and pass shape validation."""
-        manifest = {"requirements": {"02::FR-03.01": {"id": "FR-03.01"}}}
-        with pytest.raises(KeyNotDerivedFromId, match="disagrees with its node id"):
-            assert_keys_derive_from_ids(manifest)
-
-    def test_an_id_derived_key_passes(self):
-        assert_keys_derive_from_ids(
-            {"requirements": {"03::FR-03.01": {"id": "FR-03.01"}}})
-
-
-class TestIntegrityErrorsReachTheOperator:
-    """Both integrity errors must NOT be ValueErrors.
-
-    `_layer_coverage_regen` regenerates a base+head manifest inside
-    `except (OSError, ValueError)` and degrades to None, which the removal / cross-layer
-    verifiers render as the fixed string "git unavailable / no base ref / collector
-    unavailable". A ValueError subclass is therefore swallowed and reported as an
-    INFRASTRUCTURE fault -- sending an operator to check git, the base ref and the
-    collector while the real cause is a duplicate FR id in their own spec. Subclassing
-    Exception lets it reach the outer `except Exception`, which names the type."""
-
-    def test_neither_error_is_a_valueerror(self):
-        for exc in (DuplicateRequirementId, KeyNotDerivedFromId):
-            assert issubclass(exc, ManifestIntegrityError)
-            assert not issubclass(exc, ValueError), (
-                f"{exc.__name__} would be swallowed by _layer_coverage_regen's "
-                "except (OSError, ValueError) and misreported as a git/collector fault"
-            )
-
-    def test_the_regen_swallow_clause_does_not_catch_them(self):
-        # The literal clause, exercised rather than described.
-        for exc in (DuplicateRequirementId("x"), KeyNotDerivedFromId("x")):
-            try:
-                raise exc
-            except (OSError, ValueError):  # noqa: B014 - mirrors _layer_coverage_regen
-                raise AssertionError(f"{type(exc).__name__} was swallowed") from None
-            except ManifestIntegrityError:
-                pass
-
-
-class TestActiveWinsTheKeyRegardlessOfOrder:
-    """A tombstone must never displace a live row, whichever spec is discovered first."""
-
-    _ACTIVE = (
-        "# S\n\n"
-        "| ID | Requirement | Priority | Layers |\n"
-        "| --- | --- | --- | --- |\n"
-        "| FR-03.01 | Live | Must | unit |\n"
-    )
-    _REMOVED = (
-        "# S\n\n"
-        "## Removed Requirements\n\n"
-        "| ID | Requirement | Priority |\n"
-        "| --- | --- | --- |\n"
-        "| FR-03.01 | Tombstone | Must |\n"
-    )
-
-    def test_active_first_then_removed(self):
-        index = build_requirement_index([
-            (self._ACTIVE, ".shipwright/planning/01-a/spec.md"), (self._REMOVED, ".shipwright/planning/02-b/spec.md")])
-        assert index.by_key["03::FR-03.01"].is_active
-        assert [r.is_active for r in index.by_display_id["FR-03.01"]] == [True]
-
-    def test_removed_first_then_active(self):
-        index = build_requirement_index([
-            (self._REMOVED, ".shipwright/planning/02-b/spec.md"), (self._ACTIVE, ".shipwright/planning/01-a/spec.md")])
-        assert index.by_key["03::FR-03.01"].is_active
-        assert [r.is_active for r in index.by_display_id["FR-03.01"]] == [True]
+# Identity/integrity tests (duplicate-id refusal, key<->id agreement, tombstone
+# ordering) live in test_traceability_contract_integrity.py -- split there purely to
+# keep this file under the 300-LOC bloat-baseline threshold after AC-3's additive-shape
+# gate test (P3.2) was added, same subject family, not a different one.
